+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 - [**BREAKING** but deprecated] Specifying metrics with strings in `RecurrenceMatrix` and co is deprecated and will be removed in the next version. Use proper `Metric` instances like `Euclidean()`.
 - [**BREAKING** but deprecated] Giving a number as `scale` in `RecurrenceMatrix` is deprecated (because its pointless, just give `ε = ε*scale` instead).
 - New exported function `recurrence_threshold` that calculates the recurrence threshold for a given recurrence specification type.
-- RecurrenceAnalysis.jl has been updated to DynamicalSystems.jl (using StateSpaceSets.jl) and hosts its own docs
+- RecurrenceAnalysis.jl has been updated to DynamicalSystems.jl (using StateSpaceSets.jl) and hosts its own docs.
 
 ## v1.8.1
 - Better support for iteration over `AbstractRecurrenceMatrix` with functions from `Base`.